refactor: move message into it's own dir
